% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Psi.varest.R
\name{Psi.varest}
\alias{Psi.varest}
\title{Compute Psi for Proxy VAR}
\usage{
Psi.varest(x, nstep = 10, instrumented)
}
\arguments{
\item{x}{A VAR model object (class "varest").}

\item{nstep}{The number of forecast steps.}

\item{instrumented}{A character string for the instrumented variable.}
}
\value{
A 3D array representing the Psi matrix.
}
\description{
Calculates the Psi matrix for proxy VAR analysis.
}
